# openapi.yaml
openapi: "3.0.3"
info:
  title: Saukele API
  version: "1.0.0"
  description: Wedding gift management system with family tree, escrow-based pool funding, and currency snapshot locking.
servers:
  - url: http://localhost:3000
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────────────────
  # Standard responses (переиспользуемые)
  # ─────────────────────────────────────
  responses:
    BadRequest:
      description: Bad request — invalid input or business rule violation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required — missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ValidationError:
      description: Request body validation failed (Joi)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    CreatePoolRequest:
      type: object
      required: [weddingId, name, targetKzt]
      properties:
        weddingId:
          type: integer
          example: 1
        name:
          type: string
          example: "Kitchen Set"
        description:
          type: string
        targetKzt:
          type: integer
          example: 500000
        familyOnly:
          type: boolean
          default: false

        PoolResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        weddingId:
          type: integer
          example: 1
        name:
          type: string
          example: "Kitchen Set"
        description:
          type: string
          example: "contribute to their new kitchen"
        targetKzt:
          type: integer
          example: 500000
        remainingTarget:
          type: integer
          example: 350000
        totalFunded:
          type: integer
          example: 150000
        status:
          type: string
          enum: [PENDING, FUNDING, FUNDED, PURCHASED, DELIVERED]
          example: "FUNDING"
        familyOnly:
          type: boolean
          example: false
        createdAt:
          type: string
          format: date-time
          example: "2025-06-01T10:30:00.000Z"
    UpdatePoolRequest:
      type: object
      properties:
        name:
          type: string
          example: "Updated Kitchen Set"
        description:
          type: string
        targetKzt:
          type: integer
          example: 600000
        familyOnly:
          type: boolean

    # ─────────────────────────────
    # Contributions
    # ─────────────────────────────
...
    PaginatedContributions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ContributionResponse"
        pagination:
          $ref: "#/components/schemas/CursorPagination"

    # ─────────────────────────────
    # Family Tree
    # ─────────────────────────────
...
    CursorPagination:
      type: object
      properties:
        limit:
          type: integer
          example: 10
        nextCursor:
          type: integer
          nullable: true
          example: 42
        hasNextPage:
          type: boolean
          example: true
...
paths:
...
  /pools:
    post:
      tags: [Pools]
      summary: Create a new gift pool (COUPLE or SUPER_ADMIN only)
      description: First step in escrow state machine. Pool is created in PENDING status.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePoolRequest"
      responses:
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [Pools]
      summary: List all gift pools for a wedding
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: weddingId
          required: true
          schema:
            type: integer
          description: Filter by wedding ID
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, FUNDING, FUNDED, PURCHASED, DELIVERED]
          description: Filter by pool status
      responses:
        "200":
          description: Array of pools
          content:
            application/json:
              type: array
              items:
                $ref: "#/components/schemas/PoolResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /pools/{id}:
    get:
      tags: [Pools]
      summary: Get a gift pool by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [Pools]
      summary: Update a gift pool (COUPLE or SUPER_ADMIN only)
      description: Can only edit PENDING or FUNDING status pools.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePoolRequest"
      responses:
        "200":
          description: Pool updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Pools]
      summary: Delete a gift pool (COUPLE or SUPER_ADMIN only)
      description: Only pools without contributions can be deleted.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /pools/{id}/purchase:
    patch:
      tags: [Pools]
      summary: Mark a funded pool as purchased (COUPLE or SUPER_ADMIN only)
      description: "Escrow step 3: FUNDED → PURCHASED"
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool marked as purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /contributions/pool/{poolId}:
    get:
      tags: [Contributions]
      summary: List contributions for a pool (cursor-based pagination)
      security:
        - bearerAuth: []
      parameters:
        - name: poolId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: cursor
          in: query
          schema:
            type: integer
          description: Cursor from previous response (nextCursor)
      responses:
        "200":
          description: Paginated contributions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedContributions"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...